Reduced roundoff in global-to-skin transform calculation by using median.
This change is primarily to the function AnimSkin::LoadFromNif,
which, among other things, calculates the global-to-skin transform
for FO4 nif files, since this transform is not present in those files.
Unfortunately, it has been discovered that this calculation can have
quite a bit of roundoff error.  So the calculation has been changed.
Instead of using the first skin-to-bone transform in the shape to
calculate the global-to-skin transform, it uses every shape bone and
takes the median of all the calculated global-to-skin transforms.

The primary helper functions added are:
Matrix3::Transpose
CalcMedianOfFloats
CalcMedianOfVector3
CalcMedianRotation
CalcMedianMatTransform

In addition, I wrote these functions that are currently unused:
CalcAverageRotation
CalcAverageMatTransform

The resulting global-to-skin transforms definitely have less roundoff
error if the shape uses lots of bones, but there's still a noticeable
amount of roundoff error.  This is visible at the seam between eye
and face for some FO4 head models.