Comma-separated /groupbuild command; append path separator in /targetdir
The /gbuild command for BodySlide can now contain multiple group names separated by a comma.

The path for /targetdir (or /t) was missing the last backslash on Windows. Providing the backslash in the command-line used to escape a quotation mark instead if not doubled.

Additionally, the UI and the last active outfit isn't loaded anymore before building.