Cleaned up mesh coordinate transforms This commit is mostly cosmetic changes to improve clarity.  But it also fixes bugs in computing mesh::matModel.  matModel is computed in five different places, and each had at least one bug.  In fixing these bugs, I deleted the function mesh::TransformToMatrix4, which definitely had a rotation bug, and possibly a translation bug depending on how it was called.  Different callers actually needed different functionality.  So I implemented correct functionality in the five callers rather than have them call a common function that needed to be different for each of them.
I added static variables xformNifToMesh and xformMeshToNif to mesh.  These
encode, as a MatTransform, the transformation that multiplies or divides by
10, negates x, and swaps y and z.

I added member variables xformMeshToModel and xformModelToMesh to mesh.
These keep the MatTransforms corresponding to matModel and its inverse so
they can be used directly instead of having to convert nifly::Vector3 to
glm::vec4 and back.  I added setup functions for these three member variables.

I added static functions to mesh that do the same thing as VecToMeshCoords
and VecToNifCoords, but for direction vectors, difference vectors, and
distances in addition to position vectors.

I added member functions to mesh that apply xformMeshToModel and
xformModelToMesh to positions, directions, differences, and distances.

I tried my best to use these new mesh functions wherever appropriate.
I searched for all instances of '10.0f', '* 10', and '/ 10' in src to
try to catch places where VecToMeshCoords and VecToNifCoords should
have been called but weren't.

mesh::ApplyMatrix4 is no longer used anywhere, but I left it in in case
someone has a use for it someday.