Added OutfitProject::CheckShapeForDistantOrigin This new function is called from ImportOBJ, ImportFBX, and ImportNIF. It checks if the shape has an origin that's far from the center.  If it is, it asks the user if they want to move it.  The purpose of this is to reduce the precision loss when saving to a NIF with half-precision vertex coordinates.
Two things were uncertain when I wrote this:
1. What is the best test for whether the NIF has half-precision vertex
coordinates?
2. When should this function be called?
I made my best guess to both of these.

This commit is also somewhat controversial because it asks the user
to make a decision about something that's complicated enough that
they could easily make a bad choice.  Perhaps the user shouldn't
be consulted.