Implemented new smoothing brush algorithms (#437)
* Fixed some brush unwelding and normal bugs
* Fixed bug in AABBTree::AABBTreeNode::IntersectSphere that would cause it
to return a bad distance to the triangle.  This bug would make
TweakBrush::queryPoints often pick the wrong triangle as the center of its
update with the Connected-Only brush option on.  This, in turn, would cause
Mesh::ConnectedPointsInSphere to misbehave, causing brushes to unweld
vertices.  It might also sometimes cause brushes to have no effect on
some vertices that they should effect.
* Fixed bug in Mesh::ConnectedPointsInSphere that would cause it to often miss
welded vertices.  This would cause the Connected-Only brush option to
sometimes make brushes unweld vertices.
* Added Mesh::ConnectedPointsInTwoSpheres for use by brushes with the
"NeedMirrorMergedQuery" flag.  This fixes a bug that would rarely cause
points to not get adjusted.
* Fixed flaw in Mesh::SmoothNormals that would cause triangles to sometimes
flicker black.  The problem was that it would zero the normals before
calculating them.
* Fixed bug in Mesh::SmoothNormals that would cause it to produce bad
results for welded vertices in a partial update.
* Implemented new smoothing brush algorithms
* Show mirror of brush circle with x-mirror on
* Made MaskMore and MaskLess undoable.
* Made mask changes in wxGLPanel::SelectVertex undoable.
* Made InvertMask, ClearMask, OnMaskWeighted, and OnMaskBoneWeighted undoable

Co-authored-by: ousnius <ousnius@users.noreply.github.com>