Add option to disable normals calculation to "Edit" menu #515
Add option "Disable Normals Calculation" to the "Edit" menu of Outfit Studio, which will automatically turn on "Lock Normals" for all loaded meshes.

Unlike turning on "Lock Normals" manually for a shape, this new setting is only respected while working on a mesh in Outfit Studio and normals when using sliders in BodySlide are not locked.