Take gradient mask into account when conforming #513
Conforming with a mask with gradients (or lower mask values like 0.5) now keeps the old slider data and mixes it with the slider data from the new conform based on the mask factor.

This means it's now also possible to conform parts of a shape "just a bit" by applying an e.g. 20% or 50% mask to it.