Support multiple data folders when using a reference for a new project #510
Previously, when a reference was loaded that referenced external extra sliders in addition to the slider data in its local folder, it would lose the link to those when saved as a new project.

Now, both data folders are added to the reference shape in the resulting outfit project, so the link to both .osd file sources stays intact.